split the page in several sections for better maintenance
